SiteUser 엔티티, 리포, 서비스 추가, 다른 엔티티의 주키를 Long 타입으로 변경

+++ src/main/java/com/ll/exam/sbb/answer/AnswerRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import javax.transaction.Transactional;
 
-public interface AnswerRepository extends JpaRepository<Answer, Integer>, RepositoryUtil {
+public interface AnswerRepository extends JpaRepository<Answer, Long>, RepositoryUtil {
 
     @Transactional
     @Modifying
