+++ src/test/java/com/ll/exam/sbb/QuestionRepositoryTests.java
@@ -26,7 +26,7 @@ public class QuestionRepositoryTests {
         clearData();
         createSampleData();
     }
-    public static int createSampleData(QuestionRepository questionRepository) {
+    public static long createSampleData(QuestionRepository questionRepository) {
         Question q1 = new Question();
         q1.setSubject("sbb가 무엇인가요?");
         q1.setContent("sbb에 대해서 알고 싶습니다.");
@@ -42,7 +42,7 @@ public class QuestionRepositoryTests {
     }
 
     private void createSampleData() {
-        lastSampleDataId = createSampleData(questionRepository);
+        lastSampleDataId = (int) createSampleData(questionRepository);
     }
 
     public static void clearData(QuestionRepository questionRepository) {
@@ -71,18 +71,18 @@ public class QuestionRepositoryTests {
     @Test
     void 삭제() {
         assertThat(questionRepository.count()).isEqualTo(lastSampleDataId);
-        Question q = this.questionRepository.findById(1).get();
+        Question q = this.questionRepository.findById(1L).get();
         questionRepository.delete(q);
         assertThat(questionRepository.count()).isEqualTo(lastSampleDataId - 1);
     }
 
     @Test
     void 수정() {
-        Question q = this.questionRepository.findById(1).get();
+        Question q = this.questionRepository.findById(1L).get();
         q.setSubject("수정된 제목");
         questionRepository.save(q);
 
-        q = this.questionRepository.findById(1).get();
+        q = this.questionRepository.findById(1L).get();
 
         assertThat(q.getSubject()).isEqualTo("수정된 제목");
     }
